requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
websockets>=10.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pydantic>=2.0.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil>=2.8.2
PyYAML>=6.0
signalrcore>=0.9.2
aiosqlite>=0.17.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.5 